pyproject.toml
[tool.poetry]
name = "django-ultimatethumb"
version = "1.1.0"
description = "Generate thumbnails of anything."
authors = [
	"Moccu GmbH & Co. KG <info@moccu.com>",
]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/moccu/django-ultimatethumb"
repository = "https://github.com/moccu/django-ultimatethumb"
classifiers = [
	"Development Status :: 4 - Beta",
	"Environment :: Web Environment",
	"Framework :: Django",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: BSD License",
	"Operating System :: Unix",
	"Programming Language :: Python",
	"Programming Language :: Python :: 3",
]
packages = [{ include = "ultimatethumb" }]
include = ["LICENSE.rst"]

[tool.poetry.dependencies]
python = ">=3.6,<4"
importlib-metadata = {version = "*", python = "<3.8"}

Django = ">=2.2"
Pillow = ">=8.3"
command_executor = ">=0.1"
Sphinx = {version = ">=3.5", optional = true}

[tool.poetry.dev-dependencies]
pytest = ">=6.2"
pytest-cov = ">=2.12"
pytest-django = ">=4.4"
pytest-flake8 = ">=1.0"
pytest-isort = ">=2.0"
pytest-black = ">=0.3"
factory-boy = ">=3.2"
coverage = {version = ">=5.5", extras = ["toml"]}

[tool.poetry.extras]
docs = ["Sphinx"]

[build-system]
requires = ["poetry>=1.1"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --nomigrations"
testpaths = ["ultimatethumb", "tests"]
flake8-max-line-length = 96
flake8-ignore = ["E203", "E266", "E501", "W503"]
flake8-max-complexity = 18
DJANGO_SETTINGS_MODULE = "tests.settings"

[tool.black]
line-length = 96
skip-string-normalization = true
skip-numeric-underscore-normalization = true
include = "\\.pyi?$"
exclude = "/(\\.git|\\.tox|build|dist)/"

[tool.isort]
profile = "black"

[tool.coverage.run]
branch = true
source = ["ultimatethumb"]
omit = ["*/__init__.py"]

[tool.coverage.report]
exclude_lines = ["raise NotImplementedError"]
